Refactoring the redirect code into redirect maps
