+++ content/src/main/content/jcr_root/apps/acs-commons/components/utilities/redirectmappage/clientlibs/.content.xml
@@ -21,5 +21,5 @@
   -->
 <jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
     jcr:primaryType="cq:ClientLibraryFolder"
-    categories="acs-commons.redirectspage.app"
+    categories="acs-commons.redirectmappage.app"
     dependencies="[coralui2,acs-commons.angularjs,acs-commons.coralui2.angularjs]"/>
