+++ content/src/main/content/jcr_root/etc/acs-commons/redirect-maps/.content.xml
@@ -3,5 +3,5 @@
     jcr:primaryType="sling:OrderedFolder">
     <jcr:content
         jcr:primaryType="nt:unstructured"
-        jcr:title="Redirects"/>
+        jcr:title="Redirect Maps"/>
 </jcr:root>
